Refactor: split bridge.py into bridge/ subpackage
Split the bridge.py file into a bridge/ subpackage with individual
modules for each bridge implementation:

- bridge/__init__.py: re-exports + async_get_device_info()
- bridge/base.py: SamsungTVBridge ABC, mac_from_device_info,
  model_requires_encryption
- bridge/legacy.py: SamsungTVLegacyBridge
- bridge/websocket.py: SamsungTVWSBaseBridge + SamsungTVWSBridge
- bridge/encrypted.py: SamsungTVEncryptedBridge

No logic was changed — only code movement. The integration's `from
.bridge import X` imports continue to work via re-exports in
bridge/__init__.py.

Test patch paths were updated to target the correct submodules, and the
rest_api fixture was adapted to patch SamsungTVAsyncRest in both
websocket and encrypted modules.
