+++ tests/components/samsungtv/test_config_flow.py
@@ -171,7 +171,7 @@ async def test_user_legacy_does_not_ok_first_time(hass: HomeAssistant) -> None:
     """Test starting a flow by user."""
     # show form
     with patch(
-        "homeassistant.components.samsungtv.bridge.Remote",
+        "homeassistant.components.samsungtv.bridge.legacy.Remote",
         side_effect=AccessDenied("Boom"),
     ):
         result = await hass.config_entries.flow.async_init(
@@ -184,7 +184,7 @@ async def test_user_legacy_does_not_ok_first_time(hass: HomeAssistant) -> None:
             result["flow_id"], user_input=MOCK_USER_DATA
         )
 
-    with patch("homeassistant.components.samsungtv.bridge.Remote"):
+    with patch("homeassistant.components.samsungtv.bridge.legacy.Remote"):
         # entry was added
         result3 = await hass.config_entries.flow.async_configure(
             result2["flow_id"], user_input={}
@@ -207,7 +207,8 @@ async def test_user_legacy_does_not_ok_first_time(hass: HomeAssistant) -> None:
 async def test_user_websocket(hass: HomeAssistant) -> None:
     """Test starting a flow by user."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.Remote", side_effect=OSError("Boom")
+        "homeassistant.components.samsungtv.bridge.legacy.Remote",
+        side_effect=OSError("Boom"),
     ):
         # show form
         result = await hass.config_entries.flow.async_init(
@@ -287,7 +288,7 @@ async def test_user_encrypted_websocket(
 async def test_user_legacy_missing_auth(hass: HomeAssistant) -> None:
     """Test starting a flow by user with authentication."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.Remote",
+        "homeassistant.components.samsungtv.bridge.legacy.Remote",
         side_effect=AccessDenied("Boom"),
     ):
         # legacy device missing authentication
@@ -299,7 +300,7 @@ async def test_user_legacy_missing_auth(hass: HomeAssistant) -> None:
         assert result["errors"] == {"base": "auth_missing"}
 
     with patch(
-        "homeassistant.components.samsungtv.bridge.Remote",
+        "homeassistant.components.samsungtv.bridge.legacy.Remote",
         side_effect=OSError,
     ):
         # legacy device fails to connect after auth failed
@@ -314,7 +315,7 @@ async def test_user_legacy_missing_auth(hass: HomeAssistant) -> None:
 async def test_user_legacy_not_supported(hass: HomeAssistant) -> None:
     """Test starting a flow by user for not supported device."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.Remote",
+        "homeassistant.components.samsungtv.bridge.legacy.Remote",
         side_effect=UnhandledResponse("Boom"),
     ):
         # legacy device not supported
@@ -330,11 +331,11 @@ async def test_user_websocket_not_supported(hass: HomeAssistant) -> None:
     """Test starting a flow by user for not supported device."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
             side_effect=WebSocketException("Boom"),
         ),
     ):
@@ -353,11 +354,11 @@ async def test_user_websocket_access_denied(
     """Test starting a flow by user for not supported device."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
             side_effect=ConnectionClosedError(rcvd=None, sent=frames.Close(1002, "")),
         ),
     ):
@@ -375,11 +376,11 @@ async def test_user_websocket_auth_retry(hass: HomeAssistant) -> None:
     """Test starting a flow by user for not supported device."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
             side_effect=UnauthorizedError,
         ),
     ):
@@ -392,11 +393,11 @@ async def test_user_websocket_auth_retry(hass: HomeAssistant) -> None:
     assert result["errors"] == {"base": "auth_missing"}
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
         ),
     ):
         result = await hass.config_entries.flow.async_configure(
@@ -416,11 +417,11 @@ async def test_user_not_successful(hass: HomeAssistant) -> None:
     """Test starting a flow by user but no connection found."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
             side_effect=OSError("Boom"),
         ),
     ):
@@ -436,11 +437,11 @@ async def test_user_not_successful_2(hass: HomeAssistant) -> None:
     """Test starting a flow by user but no connection found."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
             side_effect=ConnectionFailure("Boom"),
         ),
     ):
@@ -537,7 +538,7 @@ async def test_ssdp_noprefix(hass: HomeAssistant) -> None:
 async def test_ssdp_legacy_missing_auth(hass: HomeAssistant) -> None:
     """Test starting a flow from discovery with authentication."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.Remote",
+        "homeassistant.components.samsungtv.bridge.legacy.Remote",
         side_effect=AccessDenied("Boom"),
     ):
         # confirm to add the entry
@@ -555,7 +556,7 @@ async def test_ssdp_legacy_missing_auth(hass: HomeAssistant) -> None:
         assert result["step_id"] == "pairing"
         assert result["errors"] == {"base": "auth_missing"}
 
-    with patch("homeassistant.components.samsungtv.bridge.Remote"):
+    with patch("homeassistant.components.samsungtv.bridge.legacy.Remote"):
         result = await hass.config_entries.flow.async_configure(
             result["flow_id"], user_input={}
         )
@@ -573,7 +574,7 @@ async def test_ssdp_legacy_missing_auth(hass: HomeAssistant) -> None:
 async def test_ssdp_legacy_not_supported(hass: HomeAssistant) -> None:
     """Test starting a flow from discovery for not supported device."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVLegacyBridge.async_try_connect",
+        "homeassistant.components.samsungtv.bridge.legacy.SamsungTVLegacyBridge.async_try_connect",
         return_value=RESULT_NOT_SUPPORTED,
     ):
         # confirm to add the entry
@@ -708,7 +709,7 @@ async def test_ssdp_encrypted_websocket_not_supported(
 ) -> None:
     """Test starting a flow from ssdp for an unsupported device populates the mac."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVEncryptedWSAsyncRemote.start_listening",
+        "homeassistant.components.samsungtv.bridge.encrypted.SamsungTVEncryptedWSAsyncRemote.start_listening",
         side_effect=WebSocketException,
     ):
         result = await hass.config_entries.flow.async_init(
@@ -725,15 +726,15 @@ async def test_ssdp_websocket_cannot_connect(hass: HomeAssistant) -> None:
     """Test starting a flow from discovery and we cannot connect."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVEncryptedWSAsyncRemote.start_listening",
+            "homeassistant.components.samsungtv.bridge.encrypted.SamsungTVEncryptedWSAsyncRemote.start_listening",
             side_effect=WebSocketException("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote",
         ) as remote_websocket,
         patch.object(remote_websocket, "open", side_effect=WebSocketException("Boom")),
     ):
@@ -765,15 +766,15 @@ async def test_ssdp_not_successful(hass: HomeAssistant) -> None:
     """Test starting a flow from discovery but no device found."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
             return_value=MOCK_DEVICE_INFO,
         ),
     ):
@@ -797,15 +798,15 @@ async def test_ssdp_not_successful_2(hass: HomeAssistant) -> None:
     """Test starting a flow from discovery but no device found."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
             side_effect=ConnectionFailure("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
             return_value=MOCK_DEVICE_INFO,
         ),
     ):
@@ -828,7 +829,7 @@ async def test_ssdp_not_successful_2(hass: HomeAssistant) -> None:
 async def test_ssdp_already_in_progress(hass: HomeAssistant) -> None:
     """Test starting a flow from discovery twice."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+        "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
         return_value=MOCK_DEVICE_INFO,
     ):
         # confirm to add the entry
@@ -850,7 +851,7 @@ async def test_ssdp_already_in_progress(hass: HomeAssistant) -> None:
 async def test_ssdp_already_configured(hass: HomeAssistant) -> None:
     """Test starting a flow from discovery when already configured."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+        "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
         return_value=MOCK_DEVICE_INFO,
     ):
         # entry was added
@@ -1122,14 +1123,14 @@ async def test_autodetect_websocket(hass: HomeAssistant) -> None:
     """Test for send key with autodetection of protocol."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote"
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote"
         ) as remote_websocket,
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVAsyncRest",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVAsyncRest",
         ) as rest_api_class,
     ):
         remote = Mock(SamsungTVWSAsyncRemote)
@@ -1174,14 +1175,14 @@ async def test_websocket_no_mac(hass: HomeAssistant, mac_address: Mock) -> None:
     mac_address.return_value = "gg:ee:tt:mm:aa:cc"
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote"
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote"
         ) as remote_websocket,
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVAsyncRest",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVAsyncRest",
         ) as rest_api_class,
     ):
         remote = Mock(SamsungTVWSAsyncRemote)
@@ -1224,7 +1225,7 @@ async def test_websocket_no_mac(hass: HomeAssistant, mac_address: Mock) -> None:
 async def test_autodetect_auth_missing(hass: HomeAssistant) -> None:
     """Test for send key with autodetection of protocol."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.Remote",
+        "homeassistant.components.samsungtv.bridge.legacy.Remote",
         side_effect=AccessDenied("Boom"),
     ) as remote:
         result = await hass.config_entries.flow.async_init(
@@ -1239,7 +1240,9 @@ async def test_autodetect_auth_missing(hass: HomeAssistant) -> None:
             call(AUTODETECT_LEGACY),
             call(AUTODETECT_LEGACY),
         ]
-    with patch("homeassistant.components.samsungtv.bridge.Remote", side_effect=OSError):
+    with patch(
+        "homeassistant.components.samsungtv.bridge.legacy.Remote", side_effect=OSError
+    ):
         result2 = await hass.config_entries.flow.async_configure(
             result["flow_id"],
             {},
@@ -1253,7 +1256,7 @@ async def test_autodetect_auth_missing(hass: HomeAssistant) -> None:
 async def test_autodetect_not_supported(hass: HomeAssistant) -> None:
     """Test for send key with autodetection of protocol."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.Remote",
+        "homeassistant.components.samsungtv.bridge.legacy.Remote",
         side_effect=[UnhandledResponse("Boom")],
     ) as remote:
         result = await hass.config_entries.flow.async_init(
@@ -1281,11 +1284,11 @@ async def test_autodetect_none(hass: HomeAssistant) -> None:
     """Test for send key with autodetection of protocol."""
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote",
             side_effect=OSError("Boom"),
         ) as remote,
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVAsyncRest.rest_device_info",
+            "homeassistant.components.samsungtv.bridge.websocket.SamsungTVAsyncRest.rest_device_info",
             side_effect=ResponseError,
         ) as rest_device_info,
     ):
@@ -1733,11 +1736,11 @@ async def test_update_legacy_missing_mac_from_dhcp_no_unique_id(
     entry.add_to_hass(hass)
     with (
         patch(
-            "homeassistant.components.samsungtv.bridge.Remote.__enter__",
+            "homeassistant.components.samsungtv.bridge.legacy.Remote.__enter__",
             return_value=True,
         ),
         patch(
-            "homeassistant.components.samsungtv.bridge.SamsungTVEncryptedWSAsyncRemote.start_listening",
+            "homeassistant.components.samsungtv.bridge.encrypted.SamsungTVEncryptedWSAsyncRemote.start_listening",
             side_effect=WebSocketException("Boom"),
         ),
     ):
@@ -1958,7 +1961,7 @@ async def test_form_reauth_websocket_not_supported(hass: HomeAssistant) -> None:
     assert result["errors"] == {}
 
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVWSAsyncRemote.open",
+        "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSAsyncRemote.open",
         side_effect=WebSocketException,
     ):
         result2 = await hass.config_entries.flow.async_configure(
@@ -2169,7 +2172,7 @@ async def test_no_update_incorrect_udn_not_matching_mac_from_dhcp(
 async def test_ssdp_update_mac(hass: HomeAssistant) -> None:
     """Ensure that MAC address is correctly updated from SSDP."""
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+        "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
         return_value=MOCK_DEVICE_INFO,
     ):
         # entry was added
@@ -2187,7 +2190,7 @@ async def test_ssdp_update_mac(hass: HomeAssistant) -> None:
     device_info = deepcopy(MOCK_DEVICE_INFO)
     device_info["device"]["wifiMac"] = "none"
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+        "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
         return_value=device_info,
     ):
         # Updated
@@ -2204,7 +2207,7 @@ async def test_ssdp_update_mac(hass: HomeAssistant) -> None:
     device_info = deepcopy(MOCK_DEVICE_INFO)
     device_info["device"]["wifiMac"] = "aa:bb:cc:dd:ee:ff"
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+        "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
         return_value=device_info,
     ):
         # Updated
@@ -2226,7 +2229,7 @@ async def test_dhcp_while_user_flow_pending(hass: HomeAssistant) -> None:
     Covers https://github.com/home-assistant/core/issues/156591.
     """
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+        "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
         return_value=None,  # Simulate device not connectable
     ):
         # Start user flow, which will show form (cannot connect)
@@ -2242,7 +2245,7 @@ async def test_dhcp_while_user_flow_pending(hass: HomeAssistant) -> None:
         ip="10.10.12.34", macaddress="aabbccddeeff", hostname="fake_hostname"
     )
     with patch(
-        "homeassistant.components.samsungtv.bridge.SamsungTVWSBridge.async_device_info",
+        "homeassistant.components.samsungtv.bridge.websocket.SamsungTVWSBridge.async_device_info",
         return_value={
             "device": {"modelName": "fake_model", "wifiMac": "aa:bb:cc:dd:ee:ff"}
         },
